Move package members module under manifest
